+++ sdk/attestation/azure-security-attestation/src/main/java/com/azure/security/attestation/AttestationClient.java
@@ -444,18 +444,12 @@ public final class AttestationClient {
      * <p>
      * Processes attestation evidence from a VBS enclave, producing an attestation result.
      * <p>The TPM attestation protocol is defined <a href='https://docs.microsoft.com/azure/attestation/virtualization-based-security-protocol'>here.</a></p>
-     * <p>Unlike OpenEnclave reports and SGX enclave quotes, TPM attestation is implemented using JSON encoded
-     * strings. </p><p>The client formats a string serialized JSON request to the service, which responds with a
-     * JSON response. The serialized JSON object exchange continues until the service responds with a JSON string
-     * with a property named {@code "report"}, whose value will be an attestation result token.</p>
      * <p><strong>Perform the first leg of a TPM attestation operation</strong></p>
      * <!-- src_embed com.azure.security.attestation.AttestationClient.attestTpm -->
      * <pre>
      * &#47;&#47; The initial payload for TPM attestation is a JSON object with a property named &quot;payload&quot;,
      * &#47;&#47; containing an object with a property named &quot;type&quot; whose value is &quot;aikcert&quot;.
      *
-     * String attestInitialPayload = &quot;&#123;&#92;&quot;payload&#92;&quot;: &#123; &#92;&quot;type&#92;&quot;: &#92;&quot;aikcert&#92;&quot; &#125; &#125;&quot;;
-     * String tpmResponse = client.attestTpm&#40;attestInitialPayload&#41;;
      * </pre>
      * <!-- end com.azure.security.attestation.AttestationClient.attestTpm -->
      *
@@ -474,18 +468,12 @@ public final class AttestationClient {
      *
      * Processes attestation evidence from a VBS enclave, producing an attestation result.
      * <p>The TPM attestation protocol is defined <a href='https://docs.microsoft.com/azure/attestation/virtualization-based-security-protocol'>here.</a></p>
-     * <p>Unlike OpenEnclave reports and SGX enclave quotes, TPM attestation is implemented using JSON encoded
-     * strings. </p><p>The client formats a string serialized JSON request to the service, which responds with a
-     * JSON response. The serialized JSON object exchange continues until the service responds with a JSON string
-     * with a property named {@code "report"}, whose value will be an attestation result token.</p>
      * <p><strong>Perform the first leg of a TPM attestation operation</strong></p>
      * <!-- src_embed com.azure.security.attestation.AttestationClient.attestTpmWithResponse -->
      * <pre>
      * &#47;&#47; The initial payload for TPM attestation is a JSON object with a property named &quot;payload&quot;,
      * &#47;&#47; containing an object with a property named &quot;type&quot; whose value is &quot;aikcert&quot;.
      *
-     * String attestInitialPayload = &quot;&#123;&#92;&quot;payload&#92;&quot;: &#123; &#92;&quot;type&#92;&quot;: &#92;&quot;aikcert&#92;&quot; &#125; &#125;&quot;;
-     * Response&lt;String&gt; tpmResponse = client.attestTpmWithResponse&#40;attestInitialPayload, Context.NONE&#41;;
      * </pre>
      * <!-- end com.azure.security.attestation.AttestationClient.attestTpmWithResponse -->
      *