+++ sdk/communication/azure-communication-callautomation/src/main/java/com/azure/communication/callautomation/implementation/models/RecordingStorageType.java
@@ -13,8 +13,8 @@ public final class RecordingStorageType extends ExpandableStringEnum<RecordingSt
     /** Static value acs for RecordingStorageType. */
     public static final RecordingStorageType ACS = fromString("acs");
 
-    /** Static value azureBlob for RecordingStorageType. */
-    public static final RecordingStorageType AZURE_BLOB = fromString("azureBlob");
+    /** Static value blobStorage for RecordingStorageType. */
+    public static final RecordingStorageType BLOB_STORAGE = fromString("blobStorage");
 
     /**
      * Creates or finds a RecordingStorageType from its string representation.
@@ -27,7 +27,11 @@ public final class RecordingStorageType extends ExpandableStringEnum<RecordingSt
         return fromString(name, RecordingStorageType.class);
     }
 
-    /** @return known RecordingStorageType values. */
+    /**
+     * Gets known RecordingStorageType values.
+     *
+     * @return known RecordingStorageType values.
+     */
     public static Collection<RecordingStorageType> values() {
         return values(RecordingStorageType.class);
     }