+++ sdk/containerregistry/azure-containers-containerregistry/src/main/java/com/azure/containers/containerregistry/ContainerRepositoryAsync.java
@@ -4,7 +4,6 @@
 package com.azure.containers.containerregistry;
 
 import com.azure.containers.containerregistry.implementation.AzureContainerRegistryImpl;
-import com.azure.containers.containerregistry.implementation.AzureContainerRegistryImplBuilder;
 import com.azure.containers.containerregistry.implementation.ContainerRegistriesImpl;
 import com.azure.containers.containerregistry.implementation.UtilsImpl;
 import com.azure.containers.containerregistry.implementation.models.RepositoryWriteableProperties;
@@ -75,15 +74,9 @@ public final class ContainerRepositoryAsync {
             throw LOGGER.logExceptionAsError(new IllegalArgumentException("'repositoryName' can't be empty."));
         }
 
-        AzureContainerRegistryImpl registryImpl = new AzureContainerRegistryImplBuilder()
-            .pipeline(httpPipeline)
-            .url(endpoint)
-            .apiVersion(version)
-            .buildClient();
-
         this.endpoint = endpoint;
         this.repositoryName = repositoryName;
-        this.serviceClient = registryImpl.getContainerRegistries();
+        this.serviceClient = new AzureContainerRegistryImpl(httpPipeline, endpoint, version).getContainerRegistries();
         this.apiVersion = version;
         this.httpPipeline = httpPipeline;