+++ sdk/kusto/azure-resourcemanager-kusto/src/main/java/com/azure/resourcemanager/kusto/models/PrincipalType.java
@@ -20,6 +20,15 @@ public final class PrincipalType extends ExpandableStringEnum<PrincipalType> {
     public static final PrincipalType USER = fromString("User");
 
     /**
+     * Creates a new instance of PrincipalType value.
+     *
+     * @deprecated Use the {@link #fromString(String)} factory method.
+     */
+    @Deprecated
+    public PrincipalType() {
+    }
+
+    /**
      * Creates or finds a PrincipalType from its string representation.
      *
      * @param name a name to look for.