+++ sdk/mysqlflexibleserver/azure-resourcemanager-mysqlflexibleserver/src/main/java/com/azure/resourcemanager/mysqlflexibleserver/models/VirtualNetworkSubnetUsageParameter.java
@@ -16,6 +16,10 @@ public final class VirtualNetworkSubnetUsageParameter {
     @JsonProperty(value = "virtualNetworkResourceId")
     private String virtualNetworkResourceId;
 
+    /** Creates an instance of VirtualNetworkSubnetUsageParameter class. */
+    public VirtualNetworkSubnetUsageParameter() {
+    }
+
     /**
      * Get the virtualNetworkResourceId property: Virtual network resource id.
      *